fix #195, add urlencode feature
Hello,

Here is a quick patch for issue #195 on Jumanji. It is tested, and do the trick.
Additionaly, it encode all non alphanumeric character using classic
url encoding, except for spaces that are replaces by '+'' char.

Not sure if it is the best way to do this, let me know.

Cheers

Signed-off-by: Sebastian Ramacher <sebastian+dev@ramacher.at>